pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "student-loan-intelligence"
version = "2.0.0"
description = "AI-powered student loan document processing and comparison platform"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [
    {name = "Student Loan Intelligence Team", email = "team@student-loan-intelligence.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial",
]
keywords = [
    "student-loans",
    "document-processing",
    "machine-learning",
    "ocr",
    "chatbot",
    "comparison",
    "financial-education"
]

dependencies = [
    # Core Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database & Storage
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "redis>=5.0.0",
    
    # HTTP & Async
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
    "asyncio-compat>=0.1.0",
    "requests>=2.31.0",
    
    # AI/ML Libraries
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "langchain>=0.0.350",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    
    # Document Processing
    "pypdfium2>=1.7.0",
    "pdfplumber>=0.10.0",
    "pillow>=10.1.0",
    "python-multipart>=0.0.6",
    
    # OCR
    "pytesseract>=0.3.10",
    "easyocr>=1.7.0",
    
    # Data Processing
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "python-dateutil>=2.8.0",
    
    # Validation & Security
    "email-validator>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Logging & Monitoring
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    
    # Utilities
    "typer>=0.9.0",
    "rich>=13.7.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",  # For testing FastAPI
    
    # Code Quality
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    
    # Development Tools
    "pre-commit>=3.5.0",
    "ipython>=8.17.0",
    "jupyter>=1.0.0",
    
    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "factory-boy>=3.3.0",
    "faker>=20.1.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-mermaid2-plugin>=1.1.0",
]

all = [
    "student-loan-intelligence[dev,test,docs]"
]

[project.urls]
Homepage = "https://github.com/your-org/student-loan-intelligence"
Documentation = "https://student-loan-intelligence.readthedocs.io/"
Repository = "https://github.com/your-org/student-loan-intelligence"
"Bug Tracker" = "https://github.com/your-org/student-loan-intelligence/issues"
Changelog = "https://github.com/your-org/student-loan-intelligence/blob/main/CHANGELOG.md"

[project.scripts]
sli = "src.main:app"
sli-api = "src.api.main:start_server"
sli-process = "src.extraction.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml", "*.txt"]

# Ruff Configuration
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugular
    "C4",  # flake8-comprehensions
    "A",   # flake8-builtins
    "YTT", # flake8-2020
    "PT",  # flake8-pytest-style
    "S",   # flake8-bandit
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "TRY", # tryceratops
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "S311",  # Standard pseudo-random generators are not suitable for cryptographic purposes
    "S101",  # Use of assert detected
    "S301",  # Pickle and modules that wrap it can be unsafe when used to deserialize untrusted data
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]

# BasedPyright Configuration
[tool.basedpyright]
typeCheckingMode = "standard"
stubPath = "typings"
reportMissingImports = "none"
reportMissingModuleSource = "none"

# MyPy Configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pypdfium2.*",
    "pdfplumber.*",
    "easyocr.*",
    "transformers.*",
    "langchain.*",
]
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit Security Linter
[tool.bandit]
exclude_dirs = ["tests"]

# Black Configuration (if used separately)
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


